aps 250211_3 auto run 수정
